# openapi.yaml
openapi: 3.1.0
info:
  title: Xiaxia Reading House V1.1 Actions
  version: 1.1.0
  description: Private reading facts and durable Xiaxia-owned reading traces. The server never generates persona text.
servers:
  - url: https://xiaxia-reading-house.onrender.com
security:
  - bearerAuth: []
paths:
...
  /api/books/{book_id}/annotations:
    get:
      operationId: listBookAnnotations
      summary: List a book's reading traces
      description: Return user annotations and, with the all filter, Xiaxia Thoughts with exact source anchors.
      parameters:
        - name: book_id
          in: path
          required: true
          description: Book UUID.
          schema:
            type: string
            format: uuid
        - name: filter
          in: query
          required: false
          description: Filter user annotations. Thoughts are included with all.
          schema:
            type: string
            enum: [all, highlights, comments, replies]
            default: all
      responses:
        "200":
          description: Filtered reading traces.
          content:
            application/json:
              schema:
                type: object
                properties:
                  filter:
                    type: string
                  annotations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Annotation"
                  xiaxia_thoughts:
                    type: array
                    items:
                      $ref: "#/components/schemas/XiaxiaThought"
                  counts:
                    type: object
                    properties:
                      annotations:
                        type: integer
                      xiaxia_thoughts:
                        type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/annotations/pending:
    get:
      operationId: listPendingAnnotations
      summary: List pending user annotations
      description: Return oldest pending user annotations, optionally limited to one book.
      parameters:
        - name: book_id
          in: query
          required: false
          description: Optional book UUID.
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          required: false
          description: Maximum result count from 1 to 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Pending annotations.
          content:
            application/json:
              schema:
                type: object
                properties:
                  annotations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Annotation"
                  count:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/annotations/{annotation_id}/seen:
    post:
      operationId: markAnnotationSeen
      summary: Mark a user annotation seen
      description: Move a pending user annotation to seen and persist Xiaxia's last annotation checkpoint.
      parameters:
        - name: annotation_id
          in: path
          required: true
          description: User annotation UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated annotation state.
          content:
            application/json:
              schema:
                type: object
                properties:
                  annotation:
                    $ref: "#/components/schemas/Annotation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/annotations/{annotation_id}/reply:
    post:
      operationId: replyToAnnotation
      summary: Create or replace Xiaxia Reply
      description: Save Xiaxia's reply to one user annotation and set its status to replied.
      parameters:
        - name: annotation_id
          in: path
          required: true
          description: User annotation UUID.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationReplyInput"
      responses:
        "200":
          description: Durable Xiaxia Reply.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    $ref: "#/components/schemas/XiaxiaReply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateAnnotationReply
      summary: Update Xiaxia Reply
      description: Modify only Xiaxia's reply while preserving the user annotation and its anchor.
      parameters:
        - name: annotation_id
          in: path
          required: true
          description: User annotation UUID.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnotationReplyInput"
      responses:
        "200":
          description: Updated Xiaxia Reply.
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    $ref: "#/components/schemas/XiaxiaReply"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteAnnotationReply
      summary: Delete Xiaxia Reply
      description: Delete only Xiaxia's reply, keep the user annotation, and return its status to seen.
      parameters:
        - name: annotation_id
          in: path
          required: true
          description: User annotation UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deletion result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  annotation_id:
                    type: string
                    format: uuid
                  annotation_status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/xiaxia/thoughts:
    get:
      operationId: listXiaxiaThoughts
      summary: List Xiaxia Thoughts
      description: List Xiaxia-owned Thoughts for one book, optionally one chapter, with stable anchors and pagination.
      parameters:
        - name: book_id
          in: query
          required: true
          description: Book UUID.
          schema:
            type: string
            format: uuid
        - name: chapter_id
          in: query
          required: false
          description: Optional chapter UUID within the book.
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          description: One-based page number.
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          required: false
          description: Results per page from 1 to 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated Xiaxia Thoughts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  thoughts:
                    type: array
                    items:
                      $ref: "#/components/schemas/XiaxiaThought"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createXiaxiaThought
      summary: Create one Xiaxia Thought
      description: Save one exact range, block, or chapter Thought. Range text must match the supplied stable block offsets.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThoughtInput"
      responses:
        "201":
          description: Created Xiaxia Thought.
          content:
            application/json:
              schema:
                type: object
                properties:
                  xiaxia_thought:
                    $ref: "#/components/schemas/XiaxiaThought"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: ACTION_API_TOKEN
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Action bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested record was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
    Anchor:
      type: object
      properties:
        scope:
          type: string
        start_block_id:
          type: [string, "null"]
        start_offset:
          type: [integer, "null"]
        end_block_id:
          type: [string, "null"]
        end_offset:
          type: [integer, "null"]
        selected_text:
          type: string
    Annotation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        chapter_id:
          type: string
          format: uuid
        selected_text:
          type: string
        start_block_id:
          type: string
        start_offset:
          type: integer
        end_block_id:
          type: string
        end_offset:
          type: integer
        comment:
          type: string
        status:
          type: string
        owner:
          type: string
        content_type:
          type: string
        xiaxia_response:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    XiaxiaThought:
      type: object
      properties:
        id:
          type: string
          format: uuid
        thought_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        chapter_id:
          type: string
          format: uuid
        scope:
          type: string
        mark_type:
          type: string
        content:
          type: string
        thought_content:
          type: string
        selected_text:
          type: string
        start_block_id:
          type: [string, "null"]
        start_offset:
          type: [integer, "null"]
        end_block_id:
          type: [string, "null"]
        end_offset:
          type: [integer, "null"]
        owner:
          type: string
        content_type:
          type: string
        user_response:
          type: [string, "null"]
        anchor:
          $ref: "#/components/schemas/Anchor"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    XiaxiaReply:
      type: object
      properties:
        id:
          type: string
          format: uuid
        annotation_id:
          type: string
          format: uuid
        response:
          type: string
        owner:
          type: string
        content_type:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AnnotationReplyInput:
      type: object
      properties:
        response:
          type: string
          minLength: 1
          maxLength: 50000
      required: [response]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        page_count:
          type: integer
        has_next:
          type: boolean
...